here some founding...
1. some pointer refreshed to 0 at beginning of func.
1. some re cal
1. some logic sample

local test result
old
BenchmarkSM2Sign-16                 3390            313408 ns/op            5426 B/op        117 allocs/op
BenchmarkSM2Sign-16                 3765            353670 ns/op            5428 B/op        117 allocs/op
BenchmarkSM2Sign-16                 3213            314760 ns/op            5448 B/op        117 allocs/op
BenchmarkSM2Sign-16                 4060            300734 ns/op            5402 B/op        117 allocs/op
BenchmarkSM2Sign-16                 3696            299816 ns/op            5431 B/op        117 allocs/op
BenchmarkSM2Verify-16                709           1637072 ns/op           78255 B/op       1602 allocs/op
BenchmarkSM2Verify-16                715           1646294 ns/op           81814 B/op       1675 allocs/op
BenchmarkSM2Verify-16                757           1653271 ns/op           78261 B/op       1617 allocs/op
BenchmarkSM2Verify-16                735           1693393 ns/op           76773 B/op       1588 allocs/op
BenchmarkSM2Verify-16                724           1634362 ns/op           79409 B/op       1643 allocs/op
BenchmarkEcdsaSign-16              49088             22950 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaSign-16              51723             22920 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaSign-16              52618             23115 ns/op            2638 B/op         32 allocs/op
BenchmarkEcdsaSign-16              52680             24454 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaSign-16              44348             27033 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaVerify-16            16728             67700 ns/op             825 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            17436             68183 ns/op             830 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            17846             68439 ns/op             825 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            17810             68123 ns/op             825 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            17280             69192 ns/op             825 B/op         16 allocs/op
PASS
ok      github.com/Hyperledger-TWGC/tjfoc-gm/sm2        30.208s

new
BenchmarkSM2Sign-16                 3735            289624 ns/op            5407 B/op        117 allocs/op
BenchmarkSM2Sign-16                 4045            277343 ns/op            5434 B/op        117 allocs/op
BenchmarkSM2Sign-16                 4028            274260 ns/op            5422 B/op        117 allocs/op
BenchmarkSM2Sign-16                 3774            277341 ns/op            5431 B/op        117 allocs/op
BenchmarkSM2Sign-16                 4213            279239 ns/op            5432 B/op        117 allocs/op
BenchmarkSM2Verify-16                775           1575745 ns/op           81951 B/op       1675 allocs/op
BenchmarkSM2Verify-16                788           1563233 ns/op           82487 B/op       1690 allocs/op
BenchmarkSM2Verify-16                717           1526143 ns/op           77000 B/op       1595 allocs/op
BenchmarkSM2Verify-16                750           1502272 ns/op           74926 B/op       1540 allocs/op
BenchmarkSM2Verify-16                782           1499108 ns/op           69170 B/op       1427 allocs/op
BenchmarkEcdsaSign-16              49248             23165 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaSign-16              51198             23449 ns/op            2638 B/op         32 allocs/op
BenchmarkEcdsaSign-16              50770             23251 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaSign-16              50355             23487 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaSign-16              51072             23238 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaVerify-16            16992             69044 ns/op             825 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            17109             69145 ns/op             830 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            16335             69115 ns/op             825 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            16818             69884 ns/op             825 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            17334             69036 ns/op             825 B/op         16 allocs/op
PASS
ok      github.com/Hyperledger-TWGC/tjfoc-gm/sm2        29.583s            826 B/op         16 allocs/op